libs: Bump gplyapi to 3.5.0
* Map versionCode as Long
* Handle all cases of immutable object modification

+++ app/src/main/java/com/aurora/store/data/event/BusEvent.kt
@@ -26,7 +26,7 @@ sealed class BusEvent : Event() {
     lateinit var error: String
 
     data class Blacklisted(val packageName: String) : BusEvent()
-    data class ManualDownload(val packageName: String, val versionCode: Int) : BusEvent()
+    data class ManualDownload(val packageName: String, val versionCode: Long) : BusEvent()
 }
 
 sealed class AuthEvent : Event() {
